Consolidate AMQP message live tests from the internal folder into public
